pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xclaw-ag-tool-guard"
version = "1.0.0"
description = "Tool invocation validation for OpenClaw agents - prevents dangerous commands and policy violations"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "XClaw AgentGuard Security Team"}
]
keywords = [
    "openclaw",
    "security",
    "tool-validation",
    "command-safety"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
requires-python = ">=3.12"
dependencies = [
    "xclaw-agentguard>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/neil-njcn/xclaw-ag-tool-guard"
Repository = "https://github.com/neil-njcn/xclaw-ag-tool-guard.git"
Issues = "https://github.com/neil-njcn/xclaw-ag-tool-guard/issues"

[project.entry-points."openclaw.skills"]
tool_guard = "xclaw_ag_tool_guard:ToolGuardSkill"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
